Числовые role flags и политика IsAdmin
Роли переведены на numeric flags, добавлена policy IsAdmin.
Обновлены TokenService, AuthConfiguration и AccessControlService.
Добавлены realtime-обновления ролей/прав через SignalR.
Улучшены UDP discovery и сохранение серверов в desktop.
Исправлены проверки MessageRepository и добавлен throttling уведомлений.

+++ API/Services/Core/Auth/TokenService.cs
@@ -39,7 +39,7 @@ public sealed class TokenService : ITokenService
         };
 
         if (role.HasValue)
-            claims.Add(new Claim(ClaimTypes.Role, role.Value.ToString()));
+            claims.Add(new Claim(ClaimTypes.Role, ((int)role.Value).ToString()));
 
         var tokenDescriptor = new SecurityTokenDescriptor
         {
